Splitted access restrictions component
